requirements.txt
streamlit>=1.43.0
google-generativeai>=0.8.4
PyPDF2>=3.0.0
python-docx>=1.0.0
sentence-transformers>=3.4.0
numpy>=1.24.0
faiss-cpu>=1.7.3